+++ moabb/analysis/plotting.py
@@ -76,6 +76,57 @@ def score_plot(data, pipelines=None):
     return fig, color_dict
 
 
+def codecarbon_plot(data, order_list=None, pipelines=None, country=""):
+    """Plot code carbon consume for the results from the benchmark
+
+    Parameters
+    ----------
+    data: output of Results.to_dataframe()
+        results on datasets
+    order_list: list of str | None
+        order of pipelines to include in this plot
+    pipelines: list of str | None
+        pipelines to include in this plot
+    country: str
+        country to include in the title
+    pipelines: list of str | None
+        pipelines to include in this plot
+
+    Returns
+    -------
+    fig: Figure
+        Pyplot handle
+    """
+    data = collapse_session_scores(data)
+    unique_ids = data["dataset"].apply(_simplify_names)
+    if len(unique_ids) != len(set(unique_ids)):
+        log.warning("Dataset names are too similar, turning off name shortening")
+    else:
+        data["dataset"] = unique_ids
+
+    if pipelines is not None:
+        data = data[data.pipeline.isin(pipelines)]
+
+    data = data.rename(columns={"carbon emission": "carbon_emission"})
+
+    fig = sea.catplot(
+        kind="bar",
+        data=data,
+        x="dataset",
+        y="carbon_emission",
+        hue="pipeline",
+        palette=PIPELINE_PALETTE,
+        height=8.5,
+        hue_order=order_list,
+    ).set(title=r"$CO_2$ emission per dataset and algorithm" + country)
+    fig.set(yscale="log")
+    fig.tight_layout()
+    fig.set_ylabels(r"$CO_2$ emission (Log Scale)")
+    fig.set_xlabels("Dataset")
+
+    return fig
+
+
 def paired_plot(data, alg1, alg2):
     """Generate a figure with a paired plot